refactor(web): 페이지 조립 구조 정리 및 반응형 레이아웃 수정
