+++ apps/web/src/_pages/moim-detail/assets/img/thumbnail.png
