+++ apps/web/src/_pages/moim-detail/index.ts
@@ -1,0 +1,3 @@
+export { CompactCard } from "./ui/compact-card";
+export { InformationContainer } from "./ui/information-container";
+export { PersonnelContainer } from "./ui/personnel-container";
