+++ apps/web/src/_pages/moim-detail/ui/personnel-container.tsx
