--- apps/web/src/app/recruitment/page.tsx
@@ -1,5 +1,0 @@
-import RecruitmentPage from "@/_pages/recruitment/ui/recruitment-page";
-
-export default function Recruitment() {
-  return <RecruitmentPage />;
-}
